+++ crates/miden-objects/src/asset/fungible.rs
@@ -27,8 +27,11 @@ pub struct FungibleAsset {
 impl FungibleAsset {
     // CONSTANTS
     // --------------------------------------------------------------------------------------------
-    /// Specifies a maximum amount value for fungible assets which can be at most a 63-bit value.
-    pub const MAX_AMOUNT: u64 = (1_u64 << 63) - 1;
+    /// Specifies the maximum amount a fungible asset can represent.
+    ///
+    /// This number was chosen so that it can be represented as a positive and negative number in a
+    /// field element. See `account_delta.masm` for more details on how this number was chosen.
+    pub const MAX_AMOUNT: u64 = 2u64.pow(63) - 2u64.pow(31);
 
     /// The serialized size of a [`FungibleAsset`] in bytes.
     ///